Refactor code to keep cmd and business logic in separated packages
